requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bootstrap4==0.1.0
boto3==1.26.9
botocore==1.29.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.1.3
django-allauth==0.54.0
django-bootstrap==0.2.4
django-bootstrap4==22.3
django-crispy-forms==1.14.0
django-jazzmin==2.6.0
django-storages==1.13.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
numpy==1.24.0
oauthlib==3.2.2
pandas==1.5.2
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
